feat: added separate component for the main map
