+++ src/App.tsx
@@ -1,34 +1,5 @@
-import {
-	VectorCustomizationItem,
-	YMapLocationRequest,
-} from '@yandex/ymaps3-types';
-import {
-	YMap,
-	YMapDefaultFeaturesLayer,
-	YMapDefaultSchemeLayer,
-} from 'ymap3-components';
-
-const defaultLocation: YMapLocationRequest = {
-	center: [27.5947648, 53.9108842],
-	zoom: 18,
-};
-
-const test: VectorCustomizationItem = {
-	tags: {
-		all: ['poi'],
-	},
-	stylers: {
-		visibility: 'off',
-	},
-};
+import { MainMap } from '@/components/map/MainMap';
 
 export const App = () => {
-	return (
-		<div style={{ width: '100%', height: '100vh' }}>
-			<YMap location={defaultLocation}>
-				<YMapDefaultSchemeLayer customization={[test]} theme='dark' />
-				<YMapDefaultFeaturesLayer />
-			</YMap>
-		</div>
-	);
+	return <MainMap />;
 };
