+++ src/components/map/MainMap/styled.ts
@@ -1,0 +1,8 @@
+import styled, { css } from 'styled-components';
+
+export const MapContainer = styled.div`
+	${({ theme }) => css`
+		width: ${theme.widths.full};
+		height: ${theme.heights.fullScreen};
+	`}
+`;
